Mejoras de backend (habian errores de hace tiempo) y tests + test de …

+++ src/main/java/org/springframework/samples/petclinic/web/GrupoController.java
@@ -12,9 +12,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.samples.petclinic.model.Grupo;
 import org.springframework.samples.petclinic.model.TipoCurso;
+import org.springframework.samples.petclinic.service.AlumnoService;
 import org.springframework.samples.petclinic.service.GrupoService;
-import org.springframework.samples.petclinic.service.exceptions.BadRequestException;
-import org.springframework.samples.petclinic.service.exceptions.DuplicatedGroupNameException;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -34,10 +33,12 @@ import lombok.extern.slf4j.Slf4j;
 public class GrupoController {
 	
 	private final GrupoService grupoService;
-	
+	private final AlumnoService alumnoService;
+
 	@Autowired
-	public GrupoController(GrupoService grupoService) {
+	public GrupoController(GrupoService grupoService, AlumnoService alumnoService) {
 		this.grupoService = grupoService;
+		this.alumnoService = alumnoService;
 	}
 	
 	@GetMapping("/all")
@@ -48,19 +49,6 @@ public class GrupoController {
 		if(session != null && session.getAttribute("type") == "profesor") {
 		Set<Grupo> all =  grupoService.getAllGrupos();
 		return ResponseEntity.ok(all);
-		}else {
-			 return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); 
-		}
-	}
-	
-	@GetMapping("/{curso}") /*Este creo q no se usa*/
-	public ResponseEntity<List<Grupo>> listaGruposPorCurso(@PathVariable("curso") TipoCurso curso, HttpServletRequest request) {
-		HttpSession session = request.getSession(false);
-		log.info("Sesión iniciada como: " + session.getAttribute("type"));
-
-		if(session != null && session.getAttribute("type") == "profesor") {
-		List<Grupo> gruposCurso = grupoService.getGruposByCourse(curso);	
-		return ResponseEntity.ok(gruposCurso);
 		}else {
 			 return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); 
 		}
@@ -117,7 +105,7 @@ public class GrupoController {
 	}
 	
 	@PostMapping("/new")
-	public ResponseEntity<?> create(@Valid @RequestBody Grupo resource, BindingResult result, HttpServletRequest request) throws DuplicatedGroupNameException{
+	public ResponseEntity<?> create(@Valid @RequestBody Grupo resource, BindingResult result, HttpServletRequest request){
 		HttpSession session = request.getSession(false);
 		log.info("Sesión iniciada como: " + session.getAttribute("type"));
 		if(session != null && session.getAttribute("type") == "profesor") {
@@ -125,9 +113,14 @@ public class GrupoController {
 			log.info("Solicitando crear grupo: {}", resource);
 			if(result.hasErrors()) {
 				return new ResponseEntity<>(result.getFieldError(), HttpStatus.NON_AUTHORITATIVE_INFORMATION);
-			}else {
+			}else {				
+				if(grupoService.exists(resource.getNombreGrupo())) {
+					return new ResponseEntity<>("Grupo ya existente", HttpStatus.BAD_REQUEST);
+				}else{
+				
 				grupoService.saveGroup(resource);
 				return new ResponseEntity<>("Grupo creado correctamente", HttpStatus.CREATED);
+				}
 			}
 		}else {
 			 return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); 
@@ -136,13 +129,18 @@ public class GrupoController {
 	
 
 	@DeleteMapping("/delete/{nombreGrupo}")
-	public ResponseEntity<?> deleteGroup(@PathVariable("nombreGrupo") String nombreGrupo, HttpServletRequest request)throws BadRequestException{
+	public ResponseEntity<?> deleteGroup(@PathVariable("nombreGrupo") String nombreGrupo, HttpServletRequest request){
 		HttpSession session = request.getSession(false);
 		log.info("Sesión iniciada como: " + session.getAttribute("type"));
 		if(session != null && session.getAttribute("type") == "profesor") {
 			log.info("Solicitando borrar grupo: {}", nombreGrupo);
-			grupoService.deleteGroup(nombreGrupo);
-			return new ResponseEntity<>("Grupo eliminado correctamente", HttpStatus.OK);
+			if(alumnoService.getStudentsPerGroup(nombreGrupo).isEmpty()) {
+				grupoService.deleteGroup(nombreGrupo);
+				return new ResponseEntity<>("Grupo eliminado correctamente", HttpStatus.OK);
+			}else {
+				return new ResponseEntity<>("No se puede borrar el grupo porque tiene alumnos", HttpStatus.BAD_REQUEST);
+			}
+			
 		}else {
 			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); 
 			}
